Allowing filtering on organisation list
In many of the places we use this list we only want to show department
and agency. Enable query dicts to be parsed if present (as in data-sets)
